laid down most of the framework for event-specific parsing of the minimum needed
