+++ lib/bundles/index.js
@@ -1,6 +1,9 @@
 'use strict';
 
-var parse = require('./parser.js');
+var parse = require('./parser');
+var installNpmPackages = require('./parser/npm');
+var installBowerPackages = require('./parser/bower');
+var recacheDashboardAssets = require('./parser/dashboard');
 var events = require('events');
 var watcher = require('./watcher.js');
 var fs = require('fs');
@@ -31,14 +34,36 @@ _bundlesDir.forEach(function(bundleFolderName) {
                 log.trace('Parsed bundle %s', bundle.name);
                 _bundles.push(bundle);
             })
+            .catch(function(err) {
+                log.error(err.stack);
+            })
     );
 });
 
 // Once all the initial parse promises have been resolved, start up the bundle watcher
 Q.all(promises)
     .then(function() {
-        watcher.on('bundleChanged', function (name) {
-            parse(name)
+        watcher.on('bundleChanged', function (name, changes) {
+            changes.forEach(function(change) {
+                var bundle = exports.find(name);
+                if (!bundle) return;
+                switch (change) {
+                    case 'npm':
+                        installNpmPackages(bundle);
+                        break;
+                    case 'bower':
+                        installBowerPackages(bundle);
+                        break;
+                    case 'dashboard':
+                        recacheDashboardAssets(bundle);
+                        break;
+                    default:
+                        log.error('Unknown change type %s for bundle %s', change, name);
+                        break;
+                }
+            });
+
+            /*parse(name)
                 .then(function(bundle) {
                     if (bundle) {
                         log.info('%s was changed, and has been reloaded from disk', name);
@@ -48,7 +73,7 @@ Q.all(promises)
                         'assuming the bundle has been deleted or moved', name);
                         exports.remove(name);
                     }
-                });
+                });*/
         });
 
         exports.emit('allLoaded', exports.all());
