+++ lib/bundles/watcher.js
@@ -9,15 +9,16 @@ var emitter = new EventEmitter();
 require('string.prototype.endswith');
 
 var backoffTimer = null;
-var changedBundles = [];
+var changes = {};
 
 // Patterns for chokidar to watch or ignore
 var watchPatterns = [
-    'bundles/**',             // Watch bundles folder
-    '!**/*___jb_*___',        // Ignore temp files created by JetBrains IDEs
-    '!bundles/*/view/**',     // Ignore view folders. Might watch them in a later NodeCG version, but not now.
-    '!**/node_modules/**',    // Ignore node_modules folders
-    '!**/bower_components/**' // Ignore bower_components folders
+    'bundles/**/dashboard/**',    // Watch dashboard folders
+    'bundles/**/package.json', // Watch bundle package.json files
+    'bundles/**/bower.json',   // Watch bundle package.json files
+    '!**/*___jb_*___',         // Ignore temp files created by JetBrains IDEs
+    '!**/node_modules/**',     // Ignore node_modules folders
+    '!**/bower_components/**'  // Ignore bower_components folders
 ];
 
 var watcher = chokidar.watch(watchPatterns, {
@@ -30,8 +31,12 @@ watcher.on('add', handleChange);
 watcher.on('change', handleChange);
 watcher.on('unlink', handleChange);
 
+// TODO: Add specific event for when a bundle is deleted (determined by nodecg.json being deleted)
 // On changed/added/deleted
 function handleChange(filepath) {
+    var filename = path.basename(filepath);
+    var dirname = path.dirname(filepath);
+
     // Extract the bundle name from the filepath
     var bundleName = '';
     var res = filepath.split(path.sep);
@@ -45,16 +50,27 @@ function handleChange(filepath) {
         prevPart = part;
     });
 
+    var type = 'unknown';
+    if (filename === 'package.json' && dirname.endsWith(bundleName)) {
+        type = 'npm';
+    } else if (filename === 'bower.json' && dirname.endsWith(bundleName)) {
+        type = 'bower';
+    } else if (dirname.indexOf('dashboard') >= 0) {
+        type = 'dashboard';
+    }
+
     if (backoffTimer) {
-        log.debug('Backoff active, delaying event for change detected in', bundleName, ': ', filepath);
-        if (changedBundles.indexOf(bundleName) < 0) {
-            changedBundles.push(bundleName);
+        log.debug('Backoff active, delaying event for change detected in', bundleName, ':', filepath);
+        if (changes[bundleName]) {
+            changes[bundleName].push(type);
+        } else {
+            changes[bundleName] = [type];
         }
         resetBackoffTimer();
     } else {
-        log.debug('Change detected in', bundleName, ': ', filepath);
+        log.debug('Change detected in', bundleName, ':', filepath);
         resetBackoffTimer();
-        emitter.emit('bundleChanged', bundleName);
+        emitter.emit('bundleChanged', bundleName, [type]);
     }
 }
 
@@ -62,11 +78,12 @@ function resetBackoffTimer() {
     clearTimeout(backoffTimer);
     backoffTimer = setTimeout(function() {
         backoffTimer = null;
-        changedBundles.forEach(function(bundleName) {
+        for (var bundleName in changes) {
+            if (!changes.hasOwnProperty(bundleName)) continue;
             log.debug('Backoff finished, emitting change event for', bundleName);
-            emitter.emit('bundleChanged', bundleName);
-        });
-        changedBundles = [];
+            emitter.emit('bundleChanged', bundleName, changes[bundleName]);
+        }
+        changes = {};
     }, 500);
 }
